requirements.txt
openai>=1.0.0
playwright>=1.43.0
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
